// apps/main/src/models/v2/iam/userMembershipRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { SubjectType } from './subjectType';

/**
 * A single membership request for a user
 */
export interface UserMembershipRequest {
  /** The membership subject type. */
  subject_type: SubjectType;
  /** The membership subject identity. */
  subject: string;
  /** The membership scope if applicable composed by <resource-kind>:<resource-id>. It will be empty for organization level memberships. */
  scope?: string;
}